unbookRoom method separated into processor
