+++ api/src/main/java/com/tinqinacademy/hotel/api/operations/hotel/unbookroom/UnbookOperation.java
@@ -1,0 +1,6 @@
+package com.tinqinacademy.hotel.api.operations.hotel.unbookroom;
+
+import com.tinqinacademy.hotel.api.base.OperationProcessor;
+
+public interface UnbookOperation extends OperationProcessor<UnbookRoomInput,UnbookRoomOutput> {
+}
