+++ api/src/main/java/com/tinqinacademy/hotel/api/operations/hotel/unbookroom/UnbookRoomInput.java
@@ -1,5 +1,6 @@
 package com.tinqinacademy.hotel.api.operations.hotel.unbookroom;
 
+import com.tinqinacademy.hotel.api.base.OperationInput;
 import jakarta.validation.constraints.NotNull;
 import lombok.*;
 
@@ -9,7 +10,7 @@ import lombok.*;
 @ToString
 @NoArgsConstructor
 @AllArgsConstructor
-public class UnbookRoomInput {
+public class UnbookRoomInput implements OperationInput {
     @NotNull
     private String bookingId;
 }
