+++ api/src/main/java/com/tinqinacademy/hotel/api/operations/hotel/unbookroom/UnbookRoomOutput.java
@@ -1,5 +1,6 @@
 package com.tinqinacademy.hotel.api.operations.hotel.unbookroom;
 
+import com.tinqinacademy.hotel.api.base.OperationOutput;
 import lombok.*;
 
 @Builder(toBuilder = true)
@@ -7,6 +8,6 @@ import lombok.*;
 @Setter
 @ToString
 @NoArgsConstructor
-public class UnbookRoomOutput {
+public class UnbookRoomOutput implements OperationOutput {
 
 }
