--- api/src/main/java/com/tinqinacademy/hotel/api/services/HotelService.java
@@ -1,14 +1,0 @@
-package com.tinqinacademy.hotel.api.services;
-
-
-import com.tinqinacademy.hotel.api.operations.hotel.bookroom.BookRoomInput;
-import com.tinqinacademy.hotel.api.operations.hotel.bookroom.BookRoomOutput;
-import com.tinqinacademy.hotel.api.operations.hotel.unbookroom.UnbookRoomInput;
-import com.tinqinacademy.hotel.api.operations.hotel.unbookroom.UnbookRoomOutput;
-
-public interface HotelService {
-
-    UnbookRoomOutput unbookRoom(UnbookRoomInput input);
-
-
-}
