+++ core/src/main/java/com/tinqinacademy/hotel/core/operations/hotel/BookRoomOperationProcessor.java
@@ -1,4 +1,4 @@
-package com.tinqinacademy.hotel.core.operations.system;
+package com.tinqinacademy.hotel.core.operations.hotel;
 
 import com.tinqinacademy.hotel.api.error.Errors;
 import com.tinqinacademy.hotel.api.operations.hotel.bookroom.BookRoomInput;
